fcp optimization to asset and home
