+++ src/app/(dashboard)/assets/page.tsx
@@ -205,7 +205,7 @@ export default function AssetsTable() {
   const [sortDirection, setSortDirection] = useState<SortDirection>("asc");
 
   const itemsPerPage = 7;
-  const { user } = useAuth();
+  const { user, isLoading: authLoading } = useAuth();
   const router = useRouter();
   const userId = user?.id;
 
@@ -507,7 +507,7 @@ export default function AssetsTable() {
 
             {/* Rows */}
             <div className="space-y-3 flex-1">
-              {loading ? (
+              {loading || authLoading ? (
                 Array.from({ length: 5 }).map((_, i) => (
                   <div
                     key={i}
