+++ src/app/(dashboard)/layout.tsx
@@ -29,8 +29,9 @@ export default function DashboardLayout({
     router.replace("/multicalendar");
   }, [isAuthenticated, isLoading, pathname, router, user]);
 
-  // Show spinner while auth is initializing (instead of blank flash)
-  if (isLoading || !isAuthenticated) {
+  // Only block when we know the user is NOT authenticated.
+  // While isLoading=true we render the full shell so SSR outputs it immediately.
+  if (!isLoading && !isAuthenticated) {
     return (
       <div className="min-h-screen flex items-center justify-center bg-[var(--page-bg)]">
         <Loader2
