+++ src/app/(dashboard)/subcontractors/page.tsx
@@ -98,7 +98,7 @@ export default function SubcontractorsPage() {
   const [sortDirection, setSortDirection] = useState<SortDirection>("asc");
 
   const itemsPerPage = 7;
-  const { user } = useAuth();
+  const { user, isLoading: authLoading } = useAuth();
   const router = useRouter();
   const userId = user?.id;
 
@@ -355,7 +355,7 @@ export default function SubcontractorsPage() {
 
             {/* Rows */}
             <div className="space-y-3 flex-1">
-              {loading ? (
+              {loading || authLoading ? (
                 Array.from({ length: 5 }).map((_, i) => (
                   <div
                     key={i}
